Use munit in FirstSuite
